+++ packages/neuron-ui/src/services/subjects.ts
@@ -1,6 +1,6 @@
 const FallbackSubject = {
   subscribe: (args: any) => {
-    console.warn('remote is not supported')
+    console.warn('The remote module is not found, please make sure the UI is running inside the Electron App')
     console.info(JSON.stringify(args))
     return {
       unsubscribe: () => {