+++ packages/neuron-wallet/src/services/sync/get-blocks.ts
@@ -3,18 +3,23 @@ import { generateCore } from 'services/sdk-core'
 
 import { Block, BlockHeader } from 'types/cell-types'
 import TypeConvert from 'types/type-convert'
-import CheckAndSave from './check-and-save'
 import Utils from './utils'
 import HexUtils from 'utils/hex'
+import CheckTx from 'services/sync/check-and-save/tx'
+import { TransactionPersistor } from 'services/tx'
+import LockUtils from 'models/lock-utils'
+import { addressesUsedSubject } from './renderer-params'
 
 export default class GetBlocks {
   private retryTime: number
   private retryInterval: number
   private core: Core
+  private url: string
 
   constructor(url: string, retryTime: number = 3, retryInterval: number = 100) {
     this.retryTime = retryTime
     this.retryInterval = retryInterval
+    this.url = url
     this.core = generateCore(url)
   }
 
@@ -33,14 +38,28 @@ export default class GetBlocks {
     return tip
   }
 
-  public checkAndSave = async (blocks: Block[], lockHashes: string[]) => {
-    let checkResult: boolean[][] = []
+  public checkAndSave = async (blocks: Block[], lockHashes: string[]): Promise<void> => {
     for (const block of blocks) {
-      const checkAndSave = new CheckAndSave(block, lockHashes)
-      const result = await checkAndSave.process()
-      checkResult.push(result)
+      for (const tx of block.transactions) {
+        const checkTx = new CheckTx(tx, this.url)
+        const addresses = await checkTx.check(lockHashes)
+        if (addresses.length > 0) {
+          for (const input of tx.inputs!) {
+            const previousTxWithStatus = await this.getTransaction(input.previousOutput!.txHash)
+            const previousTx = TypeConvert.toTransaction(previousTxWithStatus.transaction)
+            const previousOutput = previousTx.outputs![+input.previousOutput!.index]
+            input.lock = previousOutput.lock
+            input.lockHash = LockUtils.lockScriptToHash(input.lock)
+            input.capacity = previousOutput.capacity
+          }
+          await TransactionPersistor.saveFetchTx(tx)
+          addressesUsedSubject.next({
+            addresses,
+            url: this.url,
+          })
+        }
+      }
     }
-    return checkResult
   }
 
   public retryGetBlock = async (num: string): Promise<Block> => {